Details jolis

+++ app/src/main/java/com/sixnez/model/ActeurDetailledDTO.kt
@@ -2,7 +2,7 @@ package com.sixnez.model
 
 class ActeurDetailledDTO(var id: String,
                          var nomPrenom: String?,
-                         var naissance: Short?,
-                         var mort: Short?,
+                         var naissance: Int?,
+                         var mort: Int?,
                          var metier: List<String>?,
                          var filmsDTO: List<FilmActeurDTO>?)
