+++ app/src/main/java/com/sixnez/model/FilmDTO.kt
@@ -3,5 +3,5 @@ package com.sixnez.model
 class FilmDTO(var name: String,
               var imgURL: String?,
               var id: String,
-              var annee: Short,
+              var annee: Int,
               var fav: Boolean)
