+++ app/src/main/java/com/sixnez/model/FilmDetailledDTO.kt
@@ -2,7 +2,7 @@ package com.sixnez.model
 
 class FilmDetailledDTO(var title: String,
                        var imgURL: String?,
-                       var annee: Short,
+                       var annee: Int,
                        var acteurs: List<ActeurFilmDTO>?,
                        var categories: List<String>?,
                        var fav: Boolean)
